fix(bench): bind axis-A IDE filename from the canonical id
Leave identity.filename unset on the Vue IDE companion request so the host binds the canonical spelling. Label the Svelte fence as stateless by construction. Share the typed Vue runtime request with the multi-thread worker.
