+++ packages/benchmark/src/perf/axis-a-child.spec.ts
@@ -260,12 +260,14 @@ describe("axis-A perf-gate metrics are gated ONLY on empirically-present audit s
 
 // ── A MISSING IDE carrier is UNAVAILABLE, never a coerced empty-string hash ──────
 // The content-correctness pass reads each compiled carrier's IDE code + source-map
-// from the typed compile request. A missing carrier (no ideCompanion product, or
-// an absent/empty code or source-map) MUST surface carrierContentHash as null —
-// NEVER hashed from "", which would let two both-sides-missing runs compare as an
-// EQUAL empty hash and slip past the content-equality rail. These specs inject a
-// scripted host via the VerterHostFactory seam, so they need no native build (and
-// run on the unit-spec job).
+// from the typed compile request. An absent/empty code or source-map MUST surface
+// carrierContentHash as null — NEVER hashed from "", which would let two
+// both-sides-missing runs compare as an EQUAL empty hash and slip past the
+// content-equality rail. The FakeHost `products: []` arm below is a synthetic
+// missing-product shape for this unit rail; the real singular compileRequest
+// route never returns that: an admitted request either publishes every requested
+// product or throws. These specs inject a scripted host via the VerterHostFactory
+// seam, so they need no native build (and run on the unit-spec job).
 describe("axis-A carrier-content rail fails closed on a MISSING IDE carrier", () => {
   class FakeHost implements VerterHostApi {
     constructor(
@@ -282,6 +284,8 @@ describe("axis-A carrier-content rail fails closed on a MISSING IDE carrier", ()
       this.requests.push(request);
       const ide = this.ideFor(id.replace(/\\/g, "/"));
       if (!ide) {
+        // Synthetic missing-product shape for the unit rail. The real
+        // singular compileRequest route never returns this: it throws.
         return {
           canonicalId: id,
           diagnostics: { diagnostics: [], hasErrors: false },
@@ -406,7 +410,7 @@ describe("axis-A carrier-content rail fails closed on a MISSING IDE carrier", ()
       expect(request.framework).toBe("vue");
       expect(request.identity.isProduction).toBe(false);
       expect(request.identity.forceJs).toBe(false);
-      expect(request.identity.filename).toMatch(/\.vue$/);
+      expect(request.identity.filename ?? null).toBeNull();
       expect(request.options).toEqual({
         backend: "inferred",
         ssr: false,
